% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acheck-package.R
\docType{package}
\name{acheck-package}
\alias{acheck}
\alias{acheck-package}
\title{acheck: ACTED Data Processing Helpers}
\description{
Custom functions for organizing ACTED humanitarian data for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bokola.github.io/acheck/}
  \item Report bugs at \url{https://github.com/bokola/acheck/issues}
}

}
\author{
\strong{Maintainer}: Basil Okola \email{okolabasilowiti@gmail.com} (\href{https://orcid.org/0000-0002-1047-9700}{ORCID})

Authors:
\itemize{
  \item Basil Okola \email{okolabasilowiti@gmail.com} (\href{https://orcid.org/0000-0002-1047-9700}{ORCID})
}

}
\keyword{internal}
