% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_custom_total.R
\name{add_custom_total}
\alias{add_custom_total}
\title{calculate column-wise totals}
\usage{
add_custom_total(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
calculate column-wise totals
}
\examples{
\dontrun{add_custom_total()}
}
