% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{add_gt_to_word}
\alias{add_gt_to_word}
\title{Add a gtsummary objet to word doc}
\usage{
add_gt_to_word(doc, gt_obj, title, landscape = FALSE)
}
\arguments{
\item{doc}{word doc created as \code{officer::read_docx()}}

\item{gt_obj}{gtsummary object}

\item{title}{character header}

\item{landscape}{logical. page orientation}
}
\value{
a word doc
}
\description{
Add a gtsummary objet to word doc
}
\examples{
\dontrun{
add_gt_to_word(doc, gt_obj, title, landscape = FALSE)
}
}
