% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hh_disability_counts.R
\name{calculate_hh_disability_counts}
\alias{calculate_hh_disability_counts}
\title{calculate the total number of disabled individuals at the household level}
\usage{
calculate_hh_disability_counts(
  data,
  hh_id_col = "uuid",
  disability_cols = c("level_seeing_disability", "level_hearing_disability",
    "level_selfcare_disability", "level_remembering_disability",
    "level_walking_disability", "level_communicating_disability")
)
}
\arguments{
\item{data}{dataframe containing the individual disability columns and a household identifier}

\item{hh_id_col}{character name of the column storing the unique household id}

\item{disability_cols}{character vector of column names representing the domains}
}
\value{
an integer vector representing the total disabled household members for each row
}
\description{
calculate the total number of disabled individuals at the household level
}
