% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hunger_score.R
\name{calculate_hunger_score}
\alias{calculate_hunger_score}
\title{calculate household hunger score and categorical scale}
\usage{
calculate_hunger_score(
  data,
  nofood_col = "hhsnofood",
  nofood_fr_col = "hhsnofood_fr",
  noteat_col = "hhsnoteat",
  noteat_fr_col = "hhsnoteat_fr",
  bedhung_col = "hhsbedhung",
  bedhung_fr_col = "hhsbedhung_fr"
)
}
\arguments{
\item{data}{dataframe containing the food security survey responses}

\item{nofood_col}{character matching pattern for no food in house column}

\item{nofood_fr_col}{character matching pattern for frequency of no food}

\item{noteat_col}{character matching pattern for went to sleep hungry column}

\item{noteat_fr_col}{character matching pattern for frequency of went to sleep hungry}

\item{bedhung_col}{character matching pattern for went 24 hours without eating column}

\item{bedhung_fr_col}{character matching pattern for frequency of 24 hours without eating}
}
\value{
an updated dataframe with columns 'hhs_score' and 'hhs_categorical'
}
\description{
calculate household hunger score and categorical scale
}
\examples{
# 
# create dummy dataset
survey_df <- data.frame(
  hhsnofood = c("yes", "no", "yes"),
  hhsnofood_fr = c("sometimes", NA, "often"),
  hhsnoteat = c("no", "yes", "yes"),
  hhsnoteat_fr = c(NA, "rarely", "often"),
  hhsbedhung = c("no", "no", "yes"),
  hhsbedhung_fr = c(NA, NA, "sometimes"),
  stringsAsFactors = FALSE
)

hhs_results <- calculate_hunger_score(survey_df)
}
