% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_esr_registrations.R
\name{check_esr_registrations}
\alias{check_esr_registrations}
\title{Title}
\usage{
check_esr_registrations(data1 = df_clean, data2 = df_esr)
}
\arguments{
\item{data1}{dataframe. registration data}

\item{data2}{dataframe. ESR data}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
\dontrun{check_esr_registrations(data1 = df_clean, data2 = df_esr)}
}
