% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{choices}
\alias{choices}
\title{UNOcha survey choices}
\format{
A tibble data frame with 17 rows and 2 columns:
\describe{
\item{name}{character. question name}
\item{label}{character. question label}
\item{county}{character. county}
\item{subcounty}{character. subcounty}
\item{ward}{character. ward}
\item{sublocation}{character. sublocation}
\item{enum_egency}{character. agency}
}
}
\source{
Internal survey
}
\usage{
choices
}
\description{
UNOcha survey choices
}
\keyword{datasets}
