% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_camp.R
\name{clean_camp}
\alias{clean_camp}
\title{clean recode beneficiary camp}
\usage{
clean_camp(x, col_name = "ben_camp")
}
\arguments{
\item{x}{dataframe}

\item{col_name}{character. column with camp names}
}
\value{
dataframe
}
\description{
clean recode beneficiary camp
}
\examples{
\dontrun{clean_camp(x, col_name = "ben_camp")}
}
