% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_kobo_excel.R
\name{clean_kobo_excel}
\alias{clean_kobo_excel}
\title{clean manual kobo excel data to align with api structure}
\usage{
clean_kobo_excel(data)
}
\arguments{
\item{data}{data.frame. the raw excel data frame (ddd)}
}
\value{
a cleaned data frame matching the transformed api structure
}
\description{
clean manual kobo excel data to align with api structure
}
\examples{
\dontrun{clean_kobo_excel(ddd)}
}
