% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_phone_number.R
\name{clean_phone_number}
\alias{clean_phone_number}
\title{clean phone number by adding 254 prexif}
\usage{
clean_phone_number(x, col = c("reg_num", "telephone_number", "proxy_num"))
}
\arguments{
\item{x}{dataframe}

\item{col}{character vector of phone number columns}
}
\value{
dataframe
}
\description{
clean phone number by adding 254 prexif
}
