% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{clean string of spaces and non alphanumeric characters and make lower case}
\usage{
clean_string(x)
}
\arguments{
\item{x}{a vector of characters}
}
\value{
a vector
}
\description{
clean string of spaces and non alphanumeric characters and make lower case
}
\examples{
\dontrun{clean_String('Nachola124')}
}
