% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_username.R
\name{clean_username}
\alias{clean_username}
\title{Clean names by deleting extra spaces and convert to title case}
\usage{
clean_username(x)
}
\arguments{
\item{x}{character vector of names to clean}
}
\value{
a vector
}
\description{
Clean names by deleting extra spaces and convert to title case
}
\examples{
\dontrun{clean_username()}
}
