% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_villages.R
\name{collapse_villages}
\alias{collapse_villages}
\title{collapse villages that share targets, separating by /}
\usage{
collapse_villages(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
collapse villages that share targets, separating by /
}
\examples{
\dontrun{collapse_villages()}
}
