% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_excel_headers.R
\name{convert_excel_headers}
\alias{convert_excel_headers}
\title{convert labeled excel headers to kobo code/value headers using form schema}
\usage{
convert_excel_headers(
  target_df,
  url = Sys.getenv("kobo_url2"),
  project_id = Sys.getenv("mpca_id"),
  token = Sys.getenv("kobo_token")
)
}
\arguments{
\item{target_df}{data.frame. the raw excel data frame with text headers (df)}

\item{url}{character. Kobotoolbox url}

\item{project_id}{character. kobo project uid}

\item{token}{character. project token}
}
\value{
a data frame with clean value headers matching df_lab
}
\description{
convert labeled excel headers to kobo code/value headers using form schema
}
