% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{create a \code{gtsummary} table}
\usage{
create_summary_table(
  data,
  cols = dplyr::everything(),
  labels = list(),
  group_by = NULL,
  strata = NULL,
  percent_type = "column",
  dichotomous_as_continuous = FALSE,
  report_median = FALSE,
  label_characteristic = "Characteristic"
)
}
\arguments{
\item{data}{dataframe}

\item{cols}{columns to include}

\item{labels}{character. labels}

\item{group_by}{character. grouping variable}

\item{strata}{character. stratifying variable}

\item{percent_type}{character. percents by row or column}

\item{dichotomous_as_continuous}{logical. treat binary 0/1 columns as continuous}

\item{report_median}{logical. whether to include the median alongside the mean for continuous variables}

\item{label_characteristic}{character. custom header text for the characteristic column}
}
\value{
a flextable object
}
\description{
create a \code{gtsummary} table
}
