% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_object.R
\name{create_survey_object}
\alias{create_survey_object}
\title{create a survey design object dynamically for any complex sampling design}
\usage{
create_survey_object(
  sample_df,
  population_df,
  id_cols,
  strata_cols = NULL,
  weight_col_name = "sampling_weight"
)
}
\arguments{
\item{sample_df}{dataframe containing the final selected sample observations}

\item{population_df}{dataframe containing the full original population sampling frame}

\item{id_cols}{character vector identifying structural cluster nodes in cascading order (e.g., c("village", "hh_id"))}

\item{strata_cols}{character vector containing names of stratification columns (optional, defaults to NULL)}

\item{weight_col_name}{character string naming the output column for calculated weights}
}
\value{
a configured survey design object
}
\description{
create a survey design object dynamically for any complex sampling design
}
