% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_survey_summary_table}
\alias{create_survey_summary_table}
\title{create a \code{gtsummary} table from a survey design object}
\usage{
create_survey_summary_table(
  data,
  cols = dplyr::everything(),
  labels = list(),
  group_by = NULL,
  strata = NULL,
  percent_type = "column",
  dichotomous_as_continuous = FALSE,
  report_median = FALSE,
  label_characteristic = "Characteristic",
  round_continuous = FALSE,
  numeric_summary_type = c("mean", "sum")
)
}
\arguments{
\item{data}{a survey design object created via the survey package}

\item{cols}{columns to include}

\item{labels}{character. labels}

\item{group_by}{character. grouping variable}

\item{strata}{character. stratifying variable}

\item{percent_type}{character. percents by row or column}

\item{dichotomous_as_continuous}{logical. treat binary 0/1 and likert scale columns as continuous}

\item{report_median}{logical. whether to include the median alongside the mean for continuous variables (ignored if numeric_summary_type is "sum")}

\item{label_characteristic}{character. custom header text for the characteristic column}

\item{round_continuous}{logical. whether to round continuous summaries to whole numbers instead of decimals}

\item{numeric_summary_type}{character. summary format for numeric columns, either "mean" or "sum"}
}
\value{
a gtsummary object that renders across all formats natively
}
\description{
create a \code{gtsummary} table from a survey design object
}
