% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_list_to_word.R
\name{export_list_to_word}
\alias{export_list_to_word}
\title{Export a list of dataframes to word}
\usage{
export_list_to_word(table_list, doc_path)
}
\arguments{
\item{table_list}{list of dataframes}

\item{doc_path}{character vector of output directory}
}
\description{
Export a list of dataframes to word
}
\examples{
\dontrun{export_list_to_word()}
}
