% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_excel_appended.R
\name{export_to_excel_appended}
\alias{export_to_excel_appended}
\title{Export a sheet to existing workbook}
\usage{
export_to_excel_appended(
  list_of_objects = df_target_p,
  sheetname = "Target and Registered ESR",
  file_name = paste0("UNOCHA_Summary_Report_", Sys.Date(), ".xlsx"),
  output_dir = "output"
)
}
\arguments{
\item{list_of_objects}{list of \code{gtsummary} object}

\item{sheetname}{character. worksheet name}

\item{file_name}{character. output file name}

\item{output_dir}{character. output directory}
}
\description{
Export a sheet to existing workbook
}
\examples{
\dontrun{export_to_excel_appended()}
}
