% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_excel_formatted.R
\name{export_to_excel_formatted}
\alias{export_to_excel_formatted}
\title{Export/write gtsummary output to excel workbook}
\usage{
export_to_excel_formatted(
  list_of_objects,
  file_name = paste0("UNOCHA_Summary_Report_", Sys.Date(), ".xlsx")
)
}
\arguments{
\item{list_of_objects}{list of \code{gtsummary} object}

\item{file_name}{character. output filename and associated path}
}
\description{
Export/write gtsummary output to excel workbook
}
\examples{
\dontrun{export_to_excel_formatted()} 
}
