% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_uuid_cols.R
\name{find_uuid_cols}
\alias{find_uuid_cols}
\title{find columns containing uuid strings}
\usage{
find_uuid_cols(df, threshold = 0.8)
}
\arguments{
\item{df}{dataframe}

\item{threshold}{numeric proportion of non-NA values that must match the uuid pattern to flag the column default is 0.8}
}
\value{
character vector of column names matching the uuid pattern
}
\description{
find columns containing uuid strings
}
\examples{
\dontrun{find_uuid_cols(df)}
}
