% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_dm_by_names.R
\name{flatten_dm_by_names}
\alias{flatten_dm_by_names}
\title{flatten all tables within a dm object into a list of data frames}
\usage{
flatten_dm_by_names(dm_obj)
}
\arguments{
\item{dm_obj}{a dm data model object}
}
\value{
a named list of flattened data frames
}
\description{
flatten all tables within a dm object into a list of data frames
}
