% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{format numeric vectors using gtsummary style number labels safely}
\usage{
format_num(x, ...)
}
\arguments{
\item{x}{numeric vector to be formatted}

\item{...}{optional arguments passed along to label_style_number}
}
\value{
a character vector of formatted numbers
}
\description{
format numeric vectors using gtsummary style number labels safely
}
