% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distinct.R
\name{get_distinct}
\alias{get_distinct}
\title{get distinct beneficiaries on select variables}
\usage{
get_distinct(
  x,
  vars_to_check = c("benef_name", "ben_number"),
  phone_var = "ben_number",
  name_var = "benef_name"
)
}
\arguments{
\item{x}{dataframe}

\item{vars_to_check}{character vector of columns to check}

\item{phone_var}{character vector of phone number column}

\item{name_var}{character vector of respondent name}
}
\value{
dataframe
}
\description{
get distinct beneficiaries on select variables
}
\examples{
\dontrun{get_distinct()}
}
