% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dupes_with_name.R
\name{get_dupes_with_name}
\alias{get_dupes_with_name}
\title{get duplicate rows, track the checked field, and subset down to designated columns via an argument}
\usage{
get_dupes_with_name(
  data,
  col = ben_name,
  cols_select = rlang::exprs(county, enum_egency, enum_name, ben_name, ben_id_number,
    dup_col, uuid)
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{col}{the unquoted column name to check for duplicates, defaults to ben_name}

\item{cols_select}{a list of unquoted column expressions to select, defaults to a predefined set}
}
\value{
a filtered and pruned dataframe of duplicate rows
}
\description{
get duplicate rows, track the checked field, and subset down to designated columns via an argument
}
