% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicates.R
\name{get_duplicates}
\alias{get_duplicates}
\title{get duplicate beneficiaries on select variables}
\usage{
get_duplicates(
  x,
  vars_to_check = c("benef_name", "ben_number"),
  phone_var = "ben_number",
  name_var = "benef_name"
)
}
\arguments{
\item{x}{dataframe}

\item{vars_to_check}{character vector of variables to check}

\item{phone_var}{character vector. phone number variable for hard check}

\item{name_var}{character vector. name of respondent}
}
\value{
dataframe
}
\description{
get duplicate beneficiaries on select variables
}
\examples{
\dontrun{get_duplicates()}
}
