% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edit_data.R
\name{get_edit_data}
\alias{get_edit_data}
\title{Read partner registration edit files}
\usage{
get_edit_data(
  partner = "PGI",
  cols_exclude = c("village", "ben_name"),
  sheet_name = NULL
)
}
\arguments{
\item{partner}{character string of partners e.g., "PGI"}

\item{cols_exclude}{character vector of columns to exclude from edits}

\item{sheet_name}{character vector of workbook sheet name}
}
\value{
dataframe
}
\description{
Read partner registration edit files
}
\examples{
\dontrun{get_edit_data()}
}
