% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_cols.R
\name{get_group_cols}
\alias{get_group_cols}
\title{Extract column names from a data dictionary based on structural boundaries}
\usage{
get_group_cols(
  dict,
  vulnerability_df,
  vulnerability_class = "physio",
  name_col = "name"
)
}
\arguments{
\item{dict}{dataframe representing the data dictionary}

\item{vulnerability_df}{a dataframe with vulnerability categories}

\item{vulnerability_class}{character vector of vulnerability class}

\item{name_col}{string. The column containing the variable names to extract}
}
\value{
a character vector of column names
}
\description{
Extract column names from a data dictionary based on structural boundaries
}
