% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlier_income.R
\name{get_outlier_income}
\alias{get_outlier_income}
\title{Split dataset by expenditure thresholds}
\usage{
get_outlier_income(df, threshold = 50000)
}
\arguments{
\item{df}{A dataframe containing the target columns.}

\item{threshold}{The numeric ceiling limit for filtering. Default is 50000.}
}
\value{
A list containing two dataframes: "clean_data" and "extreme_data".
}
\description{
Split dataset by expenditure thresholds
}
