% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partial_matches.R
\name{get_partial_matches}
\alias{get_partial_matches}
\title{find partial matches between focal and beneficiary}
\usage{
get_partial_matches(df, focal = "focal_name", beneficiary = "benef_name")
}
\arguments{
\item{df}{dataframe}

\item{focal}{character vector of focal variable}

\item{beneficiary}{character vector of beneficiary variable}
}
\value{
a dataframe
}
\description{
find partial matches between focal and beneficiary
}
\examples{
\dontrun{get_partial_matches()}
}
