% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partial_matches.R
\name{get_partial_matches_extend}
\alias{get_partial_matches_extend}
\title{find partial matches and map the single best-matching focal name onto dataframe 2}
\usage{
get_partial_matches_extend(
  df1,
  df2,
  focal = "focal_name",
  beneficiary = "benef_name",
  min_shared_words = 2
)
}
\arguments{
\item{df1}{dataframe containing the focal variable (source of matched names)}

\item{df2}{dataframe containing the beneficiary variable (target destination)}

\item{focal}{character name of the focal variable column in df1}

\item{beneficiary}{character name of the beneficiary variable column in df2}

\item{min_shared_words}{integer minimum number of unique words that must match to trigger a connection}
}
\value{
dataframe 2 with an appended 'name_possible' column containing the highest probability match from df1
}
\description{
find partial matches and map the single best-matching focal name onto dataframe 2
}
