% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partner_beneficiaries.R
\name{get_partner_beneficiaries}
\alias{get_partner_beneficiaries}
\title{filter and deduplicate beneficiary logs by agency partner}
\usage{
get_partner_beneficiaries(
  df = df_out,
  partner = "pgi",
  columns = c("ben_name", "ben_id_number")
)
}
\arguments{
\item{df}{dataframe containing master registry logs}

\item{partner}{character keyword used to identify the target agency, defaults to 'pgi'}

\item{columns}{character vector of column names to distinct by, defaults to c('ben_name', 'ben_id_number')}
}
\value{
a deduplicated dataframe containing partner specific records
}
\description{
filter and deduplicate beneficiary logs by agency partner
}
