% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pii.R
\name{get_pii}
\alias{get_pii}
\title{Get personally identifying columns}
\usage{
get_pii(
  df = df_clean,
  vars = c("lat", "long", "name", "mobile", "gps", "phone", "enum_num$",
    "^(enum|reg|ben|proxy|sec)_.*(name|num|id)")
)
}
\arguments{
\item{df}{dataframe}

\item{vars}{string of identifying columns}
}
\value{
a list of 3 dataframes: identifying, de-identified, and original df
}
\description{
Get personally identifying columns
}
\examples{
\dontrun{get_pii()}
}
