% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_villages.R
\name{get_standard_villages}
\alias{get_standard_villages}
\title{separate villages collapsed by forward slash ("|")}
\usage{
get_standard_villages(data1 = df_partner, data2 = df_clean)
}
\arguments{
\item{data1}{dataframe. ESR/partner data}

\item{data2}{dataframe. registration data}
}
\value{
a dataframe
}
\description{
separate villages collapsed by forward slash ("|")
}
\examples{
\dontrun{get_standard_villages(data1 = df_partner, data2 = df_clean)}
}
