% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_with_xlsform.R
\name{label_with_xlsform}
\alias{label_with_xlsform}
\title{convert and label an excel export using xlsform workbook sheets}
\usage{
label_with_xlsform(data_df, xlsform_path)
}
\arguments{
\item{data_df}{data.frame. the raw exported kobo data with label headers (df)}

\item{xlsform_path}{character. path to your local xlsform .xlsx workbook}
}
\value{
a clean data frame with code headers and labeled categorical values
}
\description{
convert and label an excel export using xlsform workbook sheets
}
