% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_names.R
\name{normalize_names}
\alias{normalize_names}
\title{Normalize a Name Vector}
\usage{
normalize_names(x)
}
\arguments{
\item{x}{A character vector of names.}
}
\value{
A cleaned character vector.
}
\description{
Cleans a character vector of names by collapsing extra whitespace,
trimming leading/trailing spaces, and converting to title case.
}
