% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchwork_grid.R
\name{patchwork_grid}
\alias{patchwork_grid}
\title{patchwork a list of plots with headers for sections arranged left to right with unified axes labels}
\usage{
patchwork_grid(plot_list, section_headers, plots_per_section = 1, nrow = 1)
}
\arguments{
\item{plot_list}{list of ggplot objects}

\item{section_headers}{character vector of headers to place above sections}

\item{plots_per_section}{integer number of plots that belong under each header}

\item{nrow}{integer number of rows for the internal plot grids, defaults to 1}
}
\value{
a patchwork layout object
}
\description{
patchwork a list of plots with headers for sections arranged left to right with unified axes labels
}
