% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_gender.R
\name{plot_age_gender}
\alias{plot_age_gender}
\title{plot age distribution faceted by gender using patchwork}
\usage{
plot_age_gender(data, age_group_col, gender_col, fill_col = "grey60")
}
\arguments{
\item{data}{dataframe containing household level records}

\item{age_group_col}{character name of the column containing age categories}

\item{gender_col}{character name of the gender column}

\item{fill_col}{hex code or color string for the bar fills}
}
\value{
a patchwork layout object
}
\description{
plot age distribution faceted by gender using patchwork
}
