% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{plot bar graphs}
\usage{
plot_bar(
  data,
  x_var,
  y_var = NULL,
  fill_col = "#0066A1",
  title = "variable distribution",
  subtitle = NULL,
  x_lab = NULL,
  y_lab = "count"
)
}
\arguments{
\item{data}{dataframe}

\item{x_var}{character vector x-axis variable}

\item{y_var}{character vector y-axis variable}

\item{fill_col}{character vector of hex color code for fill}

\item{title}{character vector of title}

\item{subtitle}{character vector of subtitle}

\item{x_lab}{character vector of label for x-axis}

\item{y_lab}{character vector of label for y-axis}
}
\value{
list of plot object
}
\description{
plot bar graphs
}
\examples{
\dontrun{plot_bar()}
}
