% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_top_n.R
\name{plot_bar_top_n}
\alias{plot_bar_top_n}
\title{plot bar graphs for top n frequencies}
\usage{
plot_bar_top_n(
  data,
  x_var,
  y_var = NULL,
  type = c("count", "proportion", "mean"),
  n_top = 10,
  fill_col = "#0066A1",
  title = "top variable distribution",
  subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{x_var}{character vector x-axis variable}

\item{y_var}{character vector variable to calculate mean or proportion on}

\item{type}{character type of calculation: "count", "proportion", or "mean"}

\item{n_top}{integer maximum number of highest value categories to display}

\item{fill_col}{character vector of hex color code for fill}

\item{title}{character vector of title}

\item{subtitle}{character vector of subtitle}

\item{x_lab}{character vector of label for x-axis}

\item{y_lab}{character vector of label for y-axis}
}
\value{
a ggplot object
}
\description{
plot bar graphs for top n frequencies
}
\examples{
\dontrun{plot_bar_top_n(data, x_var = "job_type", type = "count", n_top = 10)}
}
