% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_household.R
\name{plot_household}
\alias{plot_household}
\title{plot household density by clustering household level coordinates into villages}
\usage{
plot_household(data, lat_col, lon_col, village_col, county_col = NULL)
}
\arguments{
\item{data}{dataframe containing household level records}

\item{lat_col}{character name of the latitude column}

\item{lon_col}{character name of the longitude column}

\item{village_col}{character name of the village column to group by}

\item{county_col}{character optional name of the county column for coloring}
}
\value{
a ggplot object
}
\description{
plot household density by clustering household level coordinates into villages
}
