% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_household_leaflet.R
\name{plot_household_leaflet}
\alias{plot_household_leaflet}
\title{generate an interactive leaflet map of household clusters}
\usage{
plot_household_leaflet(data, lat_col, lon_col, village_col, county_col = NULL)
}
\arguments{
\item{data}{dataframe containing household level records}

\item{lat_col}{character name of the latitude column}

\item{lon_col}{character name of the longitude column}

\item{village_col}{character name of the village column}

\item{county_col}{character optional name of the county column}
}
\value{
a leaflet map object
}
\description{
generate an interactive leaflet map of household clusters
}
