% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parts_donut.R
\name{plot_parts_donut}
\alias{plot_parts_donut}
\title{visualize parts to whole using a donut layout framework}
\usage{
plot_parts_donut(data, fill_col, palette = c("#81C784", "#FFB74D", "#E65100"))
}
\arguments{
\item{data}{dataframe}

\item{fill_col}{character. the categorical column to slice}

\item{palette}{character vector. hex color values}
}
\value{
a ggplot object
}
\description{
visualize parts to whole using a donut layout framework
}
\examples{
\dontrun{plot_parts_donut()}
}
