% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess unocha registrations}
\usage{
preprocess(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
a dataframe
}
\description{
preprocess unocha registrations
}
\examples{
\dontrun{preprocess()}
}
