% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_reg.R
\name{preprocess_reg}
\alias{preprocess_reg}
\title{preprocess unocha registrations}
\usage{
preprocess_reg(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
a dataframe
}
\description{
preprocess unocha registrations
}
\examples{
\dontrun{preprocess_reg()}

}
