% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_kobo_data.R
\name{pull_kobo_data_mpca}
\alias{pull_kobo_data_mpca}
\title{pull data from any kobo API project and align structures dynamically}
\usage{
pull_kobo_data_mpca(
  url = Sys.getenv("kobo_url2"),
  project_id = Sys.getenv("mpca_id"),
  token = Sys.getenv("kobo_token")
)
}
\arguments{
\item{url}{character. Kobotoolbox url}

\item{project_id}{character. project id}

\item{token}{character. project token}
}
\value{
a clean dataframe matching manual excel column conventions universally
}
\description{
pull data from any kobo API project and align structures dynamically
}
