% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsform_asset.R
\name{read_xlsform_asset}
\alias{read_xlsform_asset}
\title{read an xlsform file and compile it into a kobo asset metadata list}
\usage{
read_xlsform_asset(xlsform_path)
}
\arguments{
\item{xlsform_path}{character. path to your local xlsform .xlsx or .xls file}
}
\value{
a named list structured exactly like a robotoolbox kobo_asset object
}
\description{
read an xlsform file and compile it into a kobo asset metadata list
}
