% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_factor_to_character.R
\name{recode_factor_to_character}
\alias{recode_factor_to_character}
\title{recode factors to character}
\usage{
recode_factor_to_character(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
dataframe
}
\description{
recode factors to character
}
\examples{
\dontrun{recode_factor_to_character(df)}
}
