% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sample.R
\name{run_sample}
\alias{run_sample}
\title{run sampling schemes}
\usage{
run_sample(
  df_frame,
  df_sample_n,
  dir = Sys.getenv("dir_unocha"),
  type = "PDM",
  name = "FDG"
)
}
\arguments{
\item{df_frame}{dataframe of sampling frame}

\item{df_sample_n}{dataframe of sample size}

\item{dir}{charcter. of output path}

\item{type}{charcter. monitoring stage}

\item{name}{charcter. file name for otput}
}
\description{
run sampling schemes
}
\examples{
\dontrun{run_sample()}
}
