% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots_to_png.R
\name{save_plots_to_png}
\alias{save_plots_to_png}
\title{save a list of active plots into a grid using the png device}
\usage{
save_plots_to_png(
  plot_list,
  output_path,
  ncol = 2,
  width = 1600,
  height = 1200,
  res = 144
)
}
\arguments{
\item{plot_list}{list of ggplot or grid objects}

\item{output_path}{character path where the combined grid will be saved}

\item{ncol}{integer number of columns in the grid}

\item{width}{integer width of output image in pixels}

\item{height}{integer height of output image in pixels}

\item{res}{integer nominal resolution in ppi}
}
\value{
invisible null saves file directly to disk
}
\description{
save a list of active plots into a grid using the png device
}
