% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_reg.R
\name{slice_reg}
\alias{slice_reg}
\title{slice registration to match starting dates}
\usage{
slice_reg(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
slice registration to match starting dates
}
\examples{
\dontrun{slice_reg()}
}
