% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey}
\alias{survey}
\title{UNOcha survey tool}
\format{
A tibble data frame with 17 rows and 2 columns:
\describe{
\item{type}{Character one among select_one select_multiple, integer, text, note, etc}
\item{name}{character. question name}
\item{label}{character. question label}
\item{hint}{character. question hint}
\item{calculation}{character. question calculation equation}
\item{required}{character. true/false specify if question required}
\item{appearance}{character. question appearance}
\item{constraint}{character. question constraint}
\item{relevant}{character. question relevant (branching logic)}
\item{choice_filter}{character. question choice filter}
\item{constraint_message}{character. question constraint message}
\item{repeat_count}{character. question repeat count}
\item{n}{character. question row number}
}
}
\source{
Internal survey
}
\usage{
survey
}
\description{
UNOcha survey tool
}
\keyword{datasets}
