% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_plots.R
\name{theme_plots}
\alias{theme_plots}
\title{publication ready theme wrapper using cowplot with gridlines}
\usage{
theme_plots(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{numeric base font size}

\item{base_family}{character font family}
}
\value{
a ggplot2 theme object
}
\description{
publication ready theme wrapper using cowplot with gridlines
}
