% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list_to_excel.R
\name{write_list_to_excel}
\alias{write_list_to_excel}
\title{export a named list of data frames to an excel workbook}
\usage{
write_list_to_excel(df_list, file_path)
}
\arguments{
\item{df_list}{a named list of data frames}

\item{file_path}{character string for the output path}
}
\description{
export a named list of data frames to an excel workbook
}
\examples{
\dontrun{write_list_to_excel()}
}
