% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_disability_counts.R
\name{calculate_disability_counts}
\alias{calculate_disability_counts}
\title{calculate total disability counts vector}
\usage{
calculate_disability_counts(
  data,
  disability_cols = c("level_seeing_disability", "level_hearing_disability",
    "level_selfcare_disability", "level_remembering_disability",
    "level_walking_disability", "level_communicating_disability")
)
}
\arguments{
\item{data}{dataframe containing the disability columns}

\item{disability_cols}{character vector of column names representing the domains}
}
\value{
an integer vector containing the sum of matching disabilities per row
}
\description{
calculate total disability counts vector
}
