% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partner_beneficiaries.R
\name{check_disability}
\alias{check_disability}
\title{validate and recode household disability flags based on repeat group ages}
\usage{
check_disability(df_out, df_disab, uuid_col = "uuid", age = 6)
}
\arguments{
\item{df_out}{dataframe containing the main household data}

\item{df_disab}{dataframe containing the individual repeat group data}

\item{uuid_col}{character string of the ID column name in df_out (defaults to "uuid")}

\item{age}{numeric lower age limit}
}
\value{
an updated version of df_out with corrected disability flags
}
\description{
validate and recode household disability flags based on repeat group ages
}
\examples{
# lower case comments without dots or dashes
# create dummy main dataset
main_df <- data.frame(
  uuid = c("hh_1", "hh_2", "hh_3"),
  HHM_disability = c("yes", "yes", "no"),
  stringsAsFactors = FALSE
)

# create dummy repeat group dataset
repeat_df <- data.frame(
  `uuid` = c("hh_1", "hh_1", "hh_2", "hh_2", "hh_3"),
  HHM_disability = c("yes", "no", "yes", "yes", "no"),
  prot_disability_ind_age = c(1, 35, 2, 1, 40),
  stringsAsFactors = FALSE,
  check.names = FALSE
)

# run the function
output_df <- check_disability(main_df, repeat_df, uuid_col = "uuid")
}
