% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_camp.R
\name{clean_camp}
\alias{clean_camp}
\title{clean recode beneficiary camp}
\usage{
clean_camp(x, col_name = "ben_camp", partner_col = "reforgp")
}
\arguments{
\item{x}{dataframe}

\item{col_name}{character column with camp names, defaults to 'ben_camp'}

\item{partner_col}{character column with partner organization names, defaults to 'reforgp'}
}
\value{
dataframe
}
\description{
clean recode beneficiary camp
}
\examples{
\dontrun{clean_camp(x, col_name = "ben_camp", partner_col = "reforgp")}
}
