% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_listing_table.R
\name{create_listing_table}
\alias{create_listing_table}
\title{create \code{gtreg} table listing of data}
\usage{
create_listing_table(
  data,
  cols = dplyr::everything(),
  labels = list(),
  group_by = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{cols}{columns to include}

\item{labels}{list. named list or formulas containing column label mappings}

\item{group_by}{character. grouping variable}
}
\value{
a gtsummary/gtreg listing object
}
\description{
create \code{gtreg} table listing of data
}
\examples{
\dontrun{create_listing_table()}
}
