% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{create a \code{gtsummary} table}
\usage{
create_summary_table(
  data,
  cols = dplyr::everything(),
  labels = list(),
  group_by = NULL,
  strata = NULL,
  percent_type = "column"
)
}
\arguments{
\item{data}{dataframe}

\item{cols}{columns to include}

\item{labels}{character. labels}

\item{group_by}{character. grouping variable}

\item{strata}{character. stratifying variable}

\item{percent_type}{character. percents by row or column}
}
\value{
a list of gtsummary object
}
\description{
create a \code{gtsummary} table
}
\examples{
\dontrun{create_summary_table()}
}
