% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_numeric.R
\name{cut_numeric}
\alias{cut_numeric}
\title{categorize numeric cols}
\usage{
cut_numeric(x, breaks = c(-Inf, 18, Inf), labels = c("<18", ">=18"))
}
\arguments{
\item{x}{a vector of numeric column}

\item{breaks}{numeric vector of breaks}

\item{labels}{character vector of labels}
}
\value{
vector
}
\description{
categorize numeric cols
}
\examples{
cut_numeric(c(1:10), c(-Inf,5, Inf), c("<=5", ">=6"))
}
