% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pps_sample.R
\name{draw_pps_sample}
\alias{draw_pps_sample}
\title{Draw PPS Household Sample}
\usage{
draw_pps_sample(
  data,
  county_var = "county",
  village_id_var = "village",
  allocation_df,
  alloc_size_var = "sample",
  seed = as.numeric(Sys.Date())
)
}
\arguments{
\item{data}{Household-level data frame.}

\item{county_var}{County variable name.}

\item{village_id_var}{Village identifier variable name.}

\item{allocation_df}{County allocation table.}

\item{alloc_size_var}{County sample size variable.}

\item{seed}{Random seed.}
}
\value{
Sampled household data frame.
}
\description{
Draw PPS Household Sample
}
