% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vulnerability_data.R
\name{get_vulnerability_data}
\alias{get_vulnerability_data}
\title{get vulnerability criteria (physiological, nutritional etc)}
\usage{
get_vulnerability_data()
}
\value{
a dataframe
}
\description{
get vulnerability criteria (physiological, nutritional etc)
}
