% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_kobo_columns.R
\name{label_kobo_columns}
\alias{label_kobo_columns}
\title{replace data column names with question labels from a kobo asset form metadata structure}
\usage{
label_kobo_columns(data, kobo_asset, lang = NULL)
}
\arguments{
\item{data}{dataframe containing the downloaded survey submission records}

\item{kobo_asset}{a labelled dataframe fetched via robotoolbox}

\item{lang}{character optional specific language label column to use if multilingual, e.g., 'English (en)'}
}
\value{
dataframe with renamed columns
}
\description{
replace data column names with question labels from a kobo asset form metadata structure
}
