% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_faceted_donuts.R
\name{plot_faceted_donuts}
\alias{plot_faceted_donuts}
\title{visualize parts to whole using an automated faceted donut grid via patchwork}
\usage{
plot_faceted_donuts(
  data,
  fill_col,
  facet_col,
  palette = c("#388E3C", "#F57C00", "#D32F2F")
)
}
\arguments{
\item{data}{dataframe}

\item{fill_col}{character the categorical variable to slice}

\item{facet_col}{character the variable to split panels by}

\item{palette}{character vector hex color values mapping to categories}
}
\value{
a patchwork grid assembly of donut plots
}
\description{
visualize parts to whole using an automated faceted donut grid via patchwork
}
\examples{
\dontrun{plot_faceted_donuts()}
}
