% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_household.R
\name{plot_household}
\alias{plot_household}
\title{plot individual household coordinates over all kenya county borders with a compass arrow}
\usage{
plot_household(data, lat_col, lon_col, county_col = NULL)
}
\arguments{
\item{data}{dataframe containing household level records}

\item{lat_col}{character name of the latitude column}

\item{lon_col}{character name of the longitude column}

\item{county_col}{character optional name of the county column for coloring}
}
\value{
a ggplot object
}
\description{
plot individual household coordinates over all kenya county borders with a compass arrow
}
