% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_kobo_data.R
\name{pull_kobo_data}
\alias{pull_kobo_data}
\title{pull data from kobo API}
\usage{
pull_kobo_data(
  url = Sys.getenv("kobo_url"),
  project_id = Sys.getenv("unocha_id"),
  token = Sys.getenv("unocha_token")
)
}
\arguments{
\item{url}{character. Kobotoolbox url}

\item{project_id}{character. project id}

\item{token}{character. project token}
}
\value{
a list
}
\description{
pull data from kobo API
}
\examples{
\dontrun{pull_kobo_data()}
}
